+++ mopidy_json_client/__init__.py
@@ -1,17 +1,13 @@
 import logging
 
-from .client import MopidyClient, SimpleClient
-from .listener import MopidyListener, SimpleListener
-
-
 __author__ = 'Ismael Asensio (ismailof@github.com)'
-__version__ = '0.5.14'
+__version__ = '0.6.0'
+
+from .client import MopidyClient, SimpleClient
 
 __all__ = [
     'MopidyClient',
     'SimpleClient',
-    'MopidyListener',
-    'SimpleListener'
 ]