+++ mopidy_json_client/methods_2_0/__init__.py
@@ -5,12 +5,31 @@ from .library import LibraryController
 from .playlists import PlaylistsController
 from .history import HistoryController
 
-_version_ = '1.1.2'
-__all__ = [
-    'PlaybackController',
-    'TracklistController',
-    'MixerController',
-    'LibraryController',
-    'PlaylistsController',
-    'HistoryController'
+_version_ = '2.0'
+
+_all_ = [
+    PlaybackController,
+    TracklistController,
+    MixerController,
+    LibraryController,
+    PlaylistsController,
+    HistoryController
+]
+
+mopidy_eventlist = [
+   'track_playback_paused',
+   'track_playback_resumed',
+   'track_playback_started',
+   'track_playback_ended',
+   'playback_state_changed',
+   'tracklist_changed',
+   'playlists_loaded',
+   'playlist_changed',
+   'playlist_deleted',
+   'options_changed',
+   'volume_changed',
+   'mute_changed',
+   'seeked',
+   'stream_title_changed',
+   'audio_message'  # extra event for gstreamer plugins like spectrum
 ]