+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.6] - Unreleased
+### Added
+- get_client_version() method
+### Removed
+- MopidyListener are now an internal class. To listen to events use the 'bind' method
+- No dependencies on Mopidy package
+
 ## [0.5.9] - Unreleased
 ### Added
 - Retry connection if lost (parameters `retry_max` and `retry_secs`)